builder filestructure
